1. add mock data(user bank account) 2. add mock exchange rate

+++ src/main/java/com/ebanking/payment/dto/APIResponse.java
@@ -12,8 +12,8 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class APIResponse<T> {
-    BigDecimal credit;
-    BigDecimal debit;
+    Double credit;
+    Double debit;
     T response;
 
     public APIResponse(int size, List<Transaction> result) {
